+++ src/Events.cpp
@@ -1,5 +1,3 @@
-#include<cstdlib>
-
 #include"..\include\Events.h"
 #include"..\include\Player.h"
 #include"..\include\Enemy.h"
@@ -49,7 +47,7 @@ void Events::combat(Enemy::Enemy& enemy, bool surprised) {
 
                 if (choice.isChoice("attack", 1)) {
                     int damage = player.strength + randint(1, 6);  // Player's attack based on strength + a six-sided die–roll
-                    enemy.health = std::max(enemy.health - damage, 0);
+                    enemy.health = (uint16_t)std::max(enemy.health - damage, 0);
                     type("\nYou dealt ", damage, " damage to the ", enemy.name, "!\nIts health is now ", enemy.health, ".\n");
                 } else if (choice.isChoice("display stats", 2)) {
                     player.displayStats();
@@ -95,9 +93,9 @@ void Events::combat(Enemy::Enemy& enemy, bool surprised) {
                 type ("\nYou remain hidden in the shadows, evading the enemy's attack.\n");
                 shadowmeld = false;
             } else {
-                uint16_t enemyDamage = std::max(enemy.attack + randint(1, 6) - player.CON - player.defense, 0);  // Enemy's attack based on its attack stat + a six-sided die roll - player's CON and defense
+                uint16_t enemyDamage = (uint16_t)std::max(enemy.attack + randint(1, 6) - player.CON - player.defense, 0);  // Enemy's attack based on its attack stat + a six-sided die roll - player's CON and defense
                 enemyDamage = (enemyDamage > 0) ? enemyDamage : 0;
-                player.health = std::max(player.health - enemyDamage, 0);
+                player.health = (uint16_t)std::max(player.health - enemyDamage, 0);
                 type("\nThe ", enemy.name, " dealt ", enemyDamage, " damage to you!\nYour health is now ", player.health, ".\n");
             }
         }
@@ -139,7 +137,7 @@ bool checkDeath(Enemy::Enemy& enemy, Player& player) {
     return 0;
 }
 
-void levelUp(Player& player, int hitdie) {
+void levelUp(Player& player, uint16_t hitdie) {
     uint16_t levels = 0;
     while (player.exp >= player.nextLevel) {
         player.level++;
